// backend/src/modules/alerts/alert.types.ts
export interface AlertManagerPayload {
    version: string;
    groupKey: string;
    status: 'firing' | 'resolved';
    receiver: string;
    groupLabels: Record<string, string>;
    commonLabels: Record<string, string>;
    commonAnnotations: Record<string, string>;
    externalURL: string;
    alerts: Alert[];
}

export interface Alert {
    status: 'firing' | 'resolved';
    labels: Record<string, string>;
    annotations: Record<string, string>;
    startsAt: string;
    endsAt: string;
    generatorURL: string;
    fingerprint: string;
}